+++ CHANGELOG.md
@@ -25,5 +25,6 @@
 ## v0.1.4 - WIP
 - [ ] Stretch with the window dynamically
 - [x] Add planning state to diagnostics
+- [x] Add brake command to diagnostics
 - [x] Update viewport key
 - [x] Remove "target" from viewport (Removed from planning)